I think I fixed it
So there was an issue with the navbar dropdown buttons clipping outside it's box. I think there is a bug with "align-content: center;" So in order to fix that I just added a media query and tested a lot of stuff to make it work properly.

I think this will function as well with more buttons because I put the restriction based on how wide the screen is. This breaks if you resize the actual page (pressing + and - buttons) so in v2 I will have to fix this