Basic footer setup
I still need to setup responsive design for the footer but for now it looks good.
I am going to add a logo soon.

For now I will wok on the javascript because the sliding animations aren't working properly. I would consider this a big change so I am going to make a commit just in-case I screw something up and need to revert.